Update e2e.yml (#977)
This pull request includes changes to the `jobs:` section in the
`.github/workflows/e2e.yml` file to improve the end-to-end testing
workflow. The most important changes include adding a container image
for the testing environment, updating the Node.js version specification,
and modifying the installation steps for dependencies.

Improvements to the end-to-end testing workflow:

* Added a container image `mcr.microsoft.com/playwright:v1.51.1-noble`
with user options for the testing environment.
* Updated the Node.js version specification to use the latest long-term
support (LTS) version.
* Modified the installation steps to remove the `npx playwright install
--with-deps` command.